requirements.txt
asgiref==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.9.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
inflection==0.5.1
itypes==1.2.0
pytz==2021.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==3.0.1
